♻️ Move configuration parsing logic to file.go for better separation of concerns
